Stop filtering out empty editors picks.
We have a sanity test which checks if editors picks are empty on the main fronts which alerts in the event of any problems.
I think this is a valuable systems test and it failing loudly in this case is of more value than leaving stale editors picks in place without us noticing.