Update to fapi-client v4.0.6
This is a small upgrade, catching up with the recent dependency updates of
https://github.com/guardian/facia-scala-client/releases/tag/v4.0.6,
before the more extensive update in https://github.com/guardian/facia-scala-client/pull/287
is introduced.

This update has already been tested in Ophan's PromotionPoller with
https://github.com/guardian/ophan/pull/5540, successfully deployed
to Prodution.

I've switched from `fapi-client-play27` to `fapi-client-play28` - the
`editors-picks-uploader` is an AWS Lambda, it doesn't actually run
the Play framework, so specifying a higher Play version doesn't really
affect it! The `fapi-client` uses `play-json` to do its JSON-parsing
(that's all that the `-play2x` suffix indicates), so we might as well use
the latest version that's available (allowing `facia-scala-client` to drop
support for `play27` at some future point).